+++ POD_3/DAL/Models/SubscriptionDetailModel.cs
@@ -20,7 +20,7 @@ namespace POD_3.DAL.Models
         public DateTime SubscriptionEndDate { get; set; }
 
         [Range(1, 10)]
-        public int AmountPaid { get; set; }
+        public double AmountPaid { get; set; }
 
         [StringLength(50)]
         public string PaymentMode { get; set; } = null!;
